🚚 Rename `opensource` to `projects`
