+++ src/__tests__/projects/stubs.ts
