+++ src/components/pages/projects/Stats/Stat/index.tsx
