+++ src/components/pages/projects/Stats/__tests__/stubs.ts
